Fix blurred images
