+++ packages/grant-explorer/src/features/discovery/RoundCard.tsx
@@ -18,6 +18,7 @@ import RoundBanner from "./RoundBanner";
 import RoundCardStat from "./RoundCardStat";
 import { useToken } from "wagmi";
 import { getAddress } from "viem";
+import { RoundDaysLeft } from "./RoundDaysLeft";
 
 type RoundCardProps = {
   round: RoundOverview;
@@ -126,40 +127,4 @@ const RoundBadge = ({ strategyName = "" }) => {
     </Badge>
   );
 };
-
-const RoundDaysLeft = ({
-  daysLeft = 0,
-  daysLeftToApply = 0,
-  isValidRoundEndTime = true,
-}) => {
-  const days = pluralize(["day", "days"]);
-  return (
-    <div className="flex-1">
-      {daysLeftToApply > 0 && (
-        <span
-          className="text-xs w-full font-mono"
-          data-testid="apply-days-left"
-        >
-          {daysLeftToApply} {days(daysLeftToApply)} left to apply
-        </span>
-      )}
-      <p className="text-xs w-full font-mono" data-testid="days-left">
-        {isValidRoundEndTime ? (
-          <span>
-            {daysLeft} {days(daysLeft)} left in round
-          </span>
-        ) : (
-          <span>No end time</span>
-        )}
-      </p>
-    </div>
-  );
-};
-
-// If we need something more advanced or to use in another place in codebase, we can pull in a library
-const pluralize =
-  ([singular = "", plural = ""]) =>
-  (num = 0) =>
-    num ? (num === 1 ? singular : plural) : "";
-
 export default RoundCard;
